# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WhatPay API",
    "description": "Plataforma de pagos peer-to-peer integrada con WhatsApp para Chile.\n\n## Autenticación\nLa mayoría de endpoints requieren un JWT Bearer token obtenido en `POST /api/v1/users/login`.\nLos endpoints admin requieren el header `x-admin-key`.\n\n## Flujo típico\n1. Registro vía bot WhatsApp (conversacional) o `POST /api/v1/users/register`\n2. `POST /api/v1/users/login` → PIN → JWT (30 min)\n3. `POST /api/v1/topup/webpay` → recarga wallet con Transbank\n4. `POST /api/v1/payments/pay` → envía pago P2P\n5. `POST /api/v1/payments/links` → crea link de cobro para comercios\n\n## Niveles KYC\n| Nivel | Límite por tx | Límite mensual |\n|-------|--------------|----------------|\n| `BASIC` | $50.000 | $200.000 |\n| `INTERMEDIATE` | $500.000 | $2.000.000 |\n| `FULL` | $2.000.000 | Sin límite |\n\n## Seguridad\n- PIN almacenado como bcrypt hash (cost 12)\n- RUT almacenado como HMAC-SHA256 (nunca en claro)\n- Bloqueo tras 3 intentos fallidos de PIN (15 min)\n- Rate limiting por IP en todos los endpoints\n- Todas las transferencias usan SELECT FOR UPDATE (sin double-spending)\n",
    "version": "0.1.0",
    "contact": {
      "name": "Pablo Cussen",
      "url": "https://cussen.cl",
      "email": "pablo@cussen.cl"
    },
    "license": {
      "name": "Portfolio Project",
      "url": "https://cussen.cl/whatpay"
    }
  },
  "servers": [
    {
...
      "description": "Producción (GCP Cloud Run — Santiago)"
...
    {
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido en POST /api/v1/users/login"
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "API key de administrador"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Datos inválidos."
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/payment-requests": {
      "post": {
        "tags": [
          "payment-requests"
        ],
        "summary": "Solicitar pago",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Crea una solicitud de pago a otro usuario. Auto-expiración en 72h. No se puede solicitar a uno mismo.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "requesterName",
                  "requesterPhone",
                  "targetPhone",
                  "amount",
                  "description"
                ],
                "properties": {
                  "requesterName": {
                    "type": "string"
                  },
                  "requesterPhone": {
                    "type": "string"
                  },
                  "targetPhone": {
                    "type": "string"
                  },
                  "targetName": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "integer",
                    "minimum": 100
                  },
                  "description": {
                    "type": "string"
                  },
                  "expiresInHours": {
                    "type": "integer",
                    "default": 72
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Solicitud creada",
            "content": {
              "application/json": {
                "example": {
                  "request": {
                    "id": "preq_xxx",
                    "status": "pending"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
    "/api/v1/payment-requests/sent": {
      "get": {
        "tags": [
          "payment-requests"
        ],
        "summary": "Solicitudes enviadas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sent",
            "content": {
              "application/json": {
                "example": {
                  "requests": [],
                  "count": 0
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
    "/api/v1/payment-requests/received": {
      "get": {
        "tags": [
          "payment-requests"
        ],
        "summary": "Solicitudes recibidas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "phone",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Teléfono (requerido)"
          }
        ],
        "responses": {
          "200": {
            "description": "Received",
            "content": {
              "application/json": {
                "example": {
                  "requests": [],
                  "count": 0
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
    "/api/v1/payment-requests/{id}": {
      "get": {
        "tags": [
          "payment-requests"
        ],
        "summary": "Detalle de solicitud",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Request",
            "content": {
              "application/json": {
                "example": {
                  "request": {}
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "payment-requests"
        ],
        "summary": "Cancelar solicitud",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelada",
            "content": {
              "application/json": {
                "example": {
                  "message": "ok"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
    "/api/v1/payment-requests/{id}/pay": {
      "post": {
        "tags": [
          "payment-requests"
        ],
        "summary": "Pagar solicitud",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "transactionRef"
                ],
                "properties": {
                  "transactionRef": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pagada",
            "content": {
              "application/json": {
                "example": {
                  "request": {}
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
    "/api/v1/payment-requests/{id}/decline": {
      "post": {
        "tags": [
          "payment-requests"
        ],
        "summary": "Rechazar solicitud",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Rechazada",
            "content": {
              "application/json": {
                "example": {
                  "request": {}
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
...
  }
}